feature[test]: refactored following issue tests: 31, 36, 47, 48, 51, 70
